moved Ability.java to top level package
