+++ app/src/androidTest/java/com/peternaggschga/gwent/data/UnitRepositoryIntegrationTest.java
@@ -7,6 +7,8 @@ import androidx.room.Room;
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
+import com.peternaggschga.gwent.Ability;
+
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
