+++ app/src/main/java/com/peternaggschga/gwent/Ability.java
@@ -1,4 +1,4 @@
-package com.peternaggschga.gwent.data;
+package com.peternaggschga.gwent;
 
 /**
  * An enum listing the different card abilities possible.
