+++ app/src/main/java/com/peternaggschga/gwent/data/UnitDao.java
@@ -9,6 +9,8 @@ import androidx.room.Insert;
 import androidx.room.OnConflictStrategy;
 import androidx.room.Query;
 
+import com.peternaggschga.gwent.Ability;
+
 import java.util.Collection;
 import java.util.List;
