+++ app/src/main/java/com/peternaggschga/gwent/domain/damage/DamageCalculatorBuildDirector.java
@@ -2,7 +2,7 @@ package com.peternaggschga.gwent.domain.damage;
 
 import androidx.annotation.NonNull;
 
-import com.peternaggschga.gwent.data.Ability;
+import com.peternaggschga.gwent.Ability;
 import com.peternaggschga.gwent.data.UnitEntity;
 
 import java.util.Collection;
