+++ app/src/test/java/com/peternaggschga/gwent/data/UnitEntityUnitTest.java
@@ -8,6 +8,7 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.only;
 import static org.mockito.Mockito.verify;
 
+import com.peternaggschga.gwent.Ability;
 import com.peternaggschga.gwent.domain.damage.DamageCalculator;
 
 import org.junit.Test;
