+++ app/src/test/java/com/peternaggschga/gwent/domain/cases/DamageCalculatorUseCaseUnitTest.java
@@ -5,7 +5,7 @@ import static org.mockito.Mockito.when;
 
 import androidx.annotation.NonNull;
 
-import com.peternaggschga.gwent.data.Ability;
+import com.peternaggschga.gwent.Ability;
 import com.peternaggschga.gwent.data.RowType;
 import com.peternaggschga.gwent.data.UnitEntity;
 import com.peternaggschga.gwent.data.UnitRepository;
